move something from app to menu
